requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
tensorflow==2.15.0
psycopg2==2.9.9
python-dotenv==1.0.1
scikit-learn==1.5.1